# CUDA: Implemented row flattening for non-glm RoPE