@slaren 2023-07-31
Seems to improve performance under WSL even more, from `0.48 seconds per pass` to `0.36 seconds per pass`.